fix: correct docker-pull retry logic
This was originally written against exceptions, but we're calling a
process and using the exit code to throw later, so it was never raising
any exceptions at the point of our retry logic. Instead we should move
it outside and use the non-exception `retrying` based on that exit code.
